+++ pkg/ff/src/utk_file_allocate_fseek.hpp
@@ -31,10 +31,11 @@
 #include <cstdio>
 #endif
 //} J.O. 16.9.2010
+#include <stdint.h>
 
 namespace utk
 {
-  typedef unsigned long long fsize_t;
+  typedef uint64_t fsize_t;
   int file_allocate_fseek(const char* path, fsize_t size);
 }
